Split code base into separate modules Fixes #12
- Sepparated different pats of th ecode into sepparate
logcally contained modules organised in packages and sub pckages
- fixed a number of typos
