+++ .gitignore
@@ -132,7 +132,7 @@ dmypy.json
 pickled_data/combined_tables.pkl
 pickled_data/data_frames_list_disambiguated.pkl
 pickled_data/data_frames_list.pkl
-pickled_data/failiures.pkl
+pickled_data/failures.pkl
 
 visualizations/*
