+++ grao_tables_parsing.py
@@ -1,1050 +1,21 @@
 #! /usr/bin/env python3.8
 
 """## Imports"""
-import matplotlib.pyplot as plt
-import pandas as pd
-import numpy as np
-
 import argparse
-import datetime
-import requests
-import pickle
-import random
-import regex
-import time
-import enum
-import json
 import os
 
-from typing import Any, Iterable, NamedTuple, TypeVar, Callable, Sequence, List, Optional, Tuple, Generator, Dict
-from collections import defaultdict
-from urllib.parse import quote
-from functools import reduce
-from itertools import chain
-from dataclasses import dataclass, field
-
-from bs4 import BeautifulSoup
-from requests.models import Response
-from wikidataintegrator import wdi_core, wdi_login
-from joblib import Parallel, delayed
-
-"""## Type Declarations"""
-
-
-class HeaderEnum(enum.IntEnum):
-  Old = 0
-  New = 1
-
-
-class TableTypeEnum(enum.IntEnum):
-  Quarterly = 0
-  Yearly = 1
-
-
-class DataTuple(NamedTuple):
-  data: Any
-  header_type: HeaderEnum
-  table_type: TableTypeEnum
-
-
-class SettlementDataTuple(NamedTuple):
-  key: Tuple[str, str, str]
-  data: Any
-
-
-class SettlementNamesForPeriod(NamedTuple):
-  name: Tuple[str, ...]
-  start: datetime.datetime
-  end: datetime.datetime
-
-
-class MunicipalityIdentifier(NamedTuple):
-  region: str
-  municipality: str
-
-
-class SettlementInfo(NamedTuple):
-  name: str
-  permanent_residents: int
-  current_residents: int
-
-
-class FullSettlementInfo(NamedTuple):
-  region: str
-  municipality: str
-  settlement: str
-  permanent_residents: int
-  current_residents: int
-
-
-class ParsedLines(NamedTuple):
-  municipality_ids: Dict[int, MunicipalityIdentifier]
-  settlements_info: Dict[int, SettlementInfo]
-
-
-T = TypeVar('T')
-U = TypeVar('U')
-
-
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args: List[Any], **kwargs: Dict[str, Any]):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
-
-
-class RegexPatternWrapper(metaclass=Singleton):
-  def __init__(self):
-    # Building regex strings
-    cap_letter = '\p{Lu}'
-    low_letter = '\p{Ll}'
-    separator = '[\||\!]\s*'
-    number = '\d+'
-
-    self.year_group = '(\d{4})'
-    self.date_group = '(\d{2}-\d{4})'
-    self.full_date_group = '(\d{2}-\d{2}-\d{4})'
-
-    name_part = f'[\s|-]*{cap_letter}*'
-    name_part_old = f'[\.|\s|-]{cap_letter}*'
-    type_abbr = f'{cap_letter}{{1,2}}\.'
-    name = f'{cap_letter}+{name_part * 3}'
-    name_old = f'{cap_letter}+{name_part_old * 3}'
-    word = f'{low_letter}+'
-    number_group = f'{separator}({number})\s*'
-
-    self.old_reg = f'ОБЛАСТ:({name_old})'
-    self.old_mun = f'ОБЩИНА:({name_old})'
-    self.region_name_new = f'{word} ({name}) {word} ({name})'
-    self.settlement_info_quarterly = f'({type_abbr}{name})\s*{number_group * 3}'
-    self.settlement_info_yearly = f'({type_abbr}{name})\s*{number_group * 6}'
-
-
-class PickleWrapper():
-
-  directory = ''
-
-  @staticmethod
-  def configure(directory: str):
-    PickleWrapper.directory = directory
-
-  @staticmethod
-  def pickle_data(data: Any, name: str):
-    directory = PickleWrapper.directory
-
-    if not os.path.exists(directory):
-      os.makedirs(directory)
-
-    with open(f'{directory}/{name}.pkl', 'wb') as f:
-      pickle.dump(data, f)
-
-  @staticmethod
-  def load_data(name: str) -> Optional[Any]:
-    directory = PickleWrapper.directory
-    path = f'{directory}/{name}.pkl'
-
-    if not os.path.exists(path):
-      return None
-
-    with open(path, 'rb') as f:
-      return pickle.load(f)
-
-
-class Pipeline(object):
-  def __init__(self, functions: Sequence[Callable[[T], Optional[T]]]):
-    self.functions = functions
-
-  def __call__(self, value: T) -> T:
-    return reduce(apply, self.functions, value)
-
-
-@dataclass
-class Configuration(object):
-  data_configuration_path: str
-  processed_tables_path: str
-  matched_tables_path: str
-  combined_tables_path: str
-  visualizations_path: str
-  pickled_data_path: str
-  credentials_path: str
-  data: List[str] = field(init=False)
-  _extra_params: Dict[Any, Any] = field(default_factory=dict)
-
-  def __post_init__(self):
-    with open(self.data_configuration_path) as file:
-      self.data = json.load(file)
-
-  def __getitem__(self, key: Any):
-    return self._extra_params.get(key, None)
-
-  def __setitem__(self, key: Any, value: Any):
-    self._extra_params[key] = value
-
-  def process_data_configuration(self) -> List[DataTuple]:
-    output = []
-    for entry in self.data:
-      output.append(Configuration._data_tuple_from_entry(entry))
-
-    return output
-
-  @staticmethod
-  def _data_tuple_from_entry(entry: str) -> Optional[DataTuple]:
-    if regex.search(RegexPatternWrapper().date_group, entry):
-      return DataTuple(entry, HeaderEnum.New, TableTypeEnum.Quarterly)
-
-    if date := regex.search(RegexPatternWrapper().year_group, entry):
-      header_type = HeaderEnum(int(date.group(1)) > 2005)
-      return DataTuple(entry, header_type, TableTypeEnum.Yearly)
-
-
-"""## Helper Functions"""
-
-
-def apply(val: T, func: Callable[[T], U]) -> U:
-  return func(val)
-
-
-def execute_in_parallel(
-  function: Callable[[T], U],
-  data_source: Generator[T, None, None],
-  num_jobs: int = -1
-) -> Optional[List[U]]:
-  result: Optional[List[U]] = []
-
-  with Parallel(n_jobs=num_jobs) as parallel:
-    result = parallel(map(delayed(function), data_source))
-
-  return result
-
-
-def fetch_raw_data(url: str, encoding: str = 'windows-1251') -> Response:
-  headers = requests.utils.default_headers()
-  headers.update({
-      'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:52.0) Gecko/20100101 Firefox/52.0'
-  })
-
-  req: Response = requests.get(url, headers)
-  req.encoding = encoding
-
-  return req
-
-
-def fix_names(name: str) -> str:
-    new_name = name
-    prob_pos = name.find('Ь')
-
-    # Some years in the names of settlements 'Ь' was used instead of 'Ъ'
-    if prob_pos != -1:
-      o_pos = name.find('О', prob_pos)
-      if o_pos != prob_pos + 1:
-        new_name = name.replace('Ь', 'Ъ')
-
-    # Some years in the names of settlements were spelled wrong
-    # correct names were taken from https://www.nsi.bg/nrnm/index.php?f=6&ezik=bul
-    names = {
-        'БОБОВДОЛ': 'БОБОВ ДОЛ',
-        'ВЪЛЧИДОЛ': 'ВЪЛЧИ ДОЛ',
-        'КАПИТАН ПЕТКО ВОЙВО': 'КАПИТАН ПЕТКО ВОЙВОДА',
-        'ДОБРИЧКА': 'ДОБРИЧ-СЕЛСКА',
-        'ДОБРИЧ СЕЛСКА': 'ДОБРИЧ-СЕЛСКА',
-        'БЕРАИНЦИ': 'БЕРАЙНЦИ',
-        'ФЕЛТФЕБЕЛ ДЕНКОВО': 'ФЕЛДФЕБЕЛ ДЕНКОВО',
-        'УРУЧОВЦИ': 'УРУЧЕВЦИ',
-        'ПОЛИКРАЙЩЕ': 'ПОЛИКРАИЩЕ',
-        'КАМЕШИЦА': 'КАМЕЩИЦА',
-        'БОГДАНОВДОЛ': 'БОГДАНОВ ДОЛ',
-        'СИНЬО БЬРДО': 'СИНЬО БЪРДО',
-        'ЗЕЛЕН ДОЛ': 'ЗЕЛЕНДОЛ',
-        'МАРИКОСТЕНОВО': 'МАРИКОСТИНОВО',
-        'САНСТЕФАНО': 'САН-СТЕФАНО',
-        'САН СТЕФАНО': 'САН-СТЕФАНО',
-        'ПЕТРОВДОЛ': 'ПЕТРОВ ДОЛ',
-        'ЧАПАЕВО': 'ЦАРСКИ ИЗВОР',
-        'ЕЛОВДОЛ': 'ЕЛОВ ДОЛ',
-        'В. ТЪРНОВО': 'ВЕЛИКО ТЪРНОВО',
-        'В.ТЪРНОВО': 'ВЕЛИКО ТЪРНОВО',
-        'ГЕНЕРАЛ-ТОШОВО': 'ГЕНЕРАЛ ТОШЕВО',
-        'ГЕНЕРАЛ ТОШОВО': 'ГЕНЕРАЛ ТОШЕВО',
-        'ГЕНЕРАЛ-ТОШЕВО': 'ГЕНЕРАЛ ТОШЕВО',
-        'БЕДЖДЕНЕ': 'БЕДЖЕНЕ',
-        'ТАЙМИШЕ': 'ТАЙМИЩЕ',
-        'СТОЯН ЗАИМОВО': 'СТОЯН-ЗАИМОВО',
-        'ДАСКАЛ АТАНАСОВО': 'ДАСКАЛ-АТАНАСОВО',
-        'СЛАВЕИНО': 'СЛАВЕЙНО',
-        'КРАЛЕВДОЛ': 'КРАЛЕВ ДОЛ',
-        'ФЕЛДФЕБЕЛ ДЯНКОВО': 'ФЕЛДФЕБЕЛ ДЕНКОВО',
-        'ДЛЪХЧЕВО САБЛЯР': 'ДЛЪХЧЕВО-САБЛЯР',
-        'СТОЯН ЗАИМОВО': 'СТОЯН-ЗАИМОВО',
-        'ГОЛЕМ ВЪРБОВНИК': 'ГОЛЯМ ВЪРБОВНИК',
-        'ПОЛКОВНИК ЖЕЛЕЗОВО': 'ПОЛКОВНИК ЖЕЛЯЗОВО',
-        'ДОБРИЧ ГРАД': 'ДОБРИЧ',
-        'ЦАР ПЕТРОВО': 'ЦАР-ПЕТРОВО',
-        'ВЪЛЧАНДОЛ': 'ВЪЛЧАН ДОЛ',
-        'ПАНАГЮРСКИ КОЛОНИ': 'ПАНАГЮРСКИ КОЛОНИИ',
-        'ГОРСКИ ГОРЕН ТРЪМБЕ': 'ГОРСКИ ГОРЕН ТРЪМБЕШ',
-        'ГОРСКИ ДОЛЕН ТРЪМБЕ': 'ГОРСКИ ДОЛЕН ТРЪМБЕШ',
-        'ГЕНЕРАЛ-КАНТАРДЖИЕВ': 'ГЕНЕРАЛ КАНТАРДЖИЕВО',
-        'ГЕНЕРАЛ КАНТАРДЖИЕВ': 'ГЕНЕРАЛ КАНТАРДЖИЕВО',
-        'АЛЕКСАНДЪР СТАМБОЛИ': 'АЛЕКСАНДЬР СТАМБОЛИЙСКИ',
-        'ПОЛКОВНИК-ЛАМБРИНОВ': 'ПОЛКОВНИК ЛАМБРИНОВО',
-        'ПОЛКОВНИК ЛАМБРИНОВ': 'ПОЛКОВНИК ЛАМБРИНОВО',
-        'ПОЛКОВНИК-СЕРАФИМОВ': 'ПОЛКОВНИК СЕРАФИМОВО',
-        'ПОЛКОВНИК СЕРАФИМОВ': 'ПОЛКОВНИК СЕРАФИМОВО'
-    }
-
-    if new_name.find('-') != -1:
-      new_name = new_name.replace('-', ' ')
-
-    new_name = names.get(new_name, new_name)
-
-    return new_name
-
-
-"""## Settlement Disambiguation Pipeline"""
-
-
-def fetch_raw_settlement_data(settlement: SettlementDataTuple) -> SettlementDataTuple:
-  name = settlement.data
-
-  # HACK!!! used to circumvent stripping of non-letter chars from the name
-  if name.find('-') != -1:
-    name = name.split('-')[1]
-
-  encoded_name = quote(name.encode('windows-1251'))
-  data = fetch_raw_data(f'https://www.nsi.bg/nrnm/index.php?ezik=bul&f=6&name={encoded_name}&code=&kind=-1')
-  req = data
-
-  if req.status_code != 200:
-    raise ValueError
-
-  return SettlementDataTuple(settlement.key, req)
-
-
-def parse_raw_settlement_data(settlement: SettlementDataTuple) -> SettlementDataTuple:
-  req = settlement.data
-  soup = BeautifulSoup(req.text, 'lxml')
-  table = soup.find_all('table')[-4]
-
-  data: Dict[str, List[SettlementNamesForPeriod]] = defaultdict(list)
-  last_key: str = ''
-
-  oldest_record_date = datetime.datetime.strptime('31.12.1899', '%d.%m.%Y')
-  rows = table.find_all('tr')
-  for row in rows[2:]:
-    cells = row.find_all('td')
-    num_cells = len(cells)
-
-    if num_cells == 2:
-      last_key = cells[0].text
-    elif num_cells == 3:
-      dates: str = cells[2].text.split('-')
-      start = datetime.datetime.strptime(dates[0].strip(), '%d.%m.%Y')
-      end = datetime.datetime.max
-
-      if len(dates[1].strip()) > 0:
-        end = datetime.datetime.strptime(dates[1].strip(), '%d.%m.%Y')
-
-      name_tuple: Tuple[str, ...] = tuple(map(lambda s: s.strip(), cells[1].text.split(',')[::-1]))
-
-      if end > oldest_record_date and len(name_tuple) > 2:
-        data[last_key].append(
-            SettlementNamesForPeriod(
-                name_tuple,
-                start,
-                end,
-            )
-        )
-
-  return SettlementDataTuple(settlement.key, data)
-
-
-def mach_key_with_code(settlement: SettlementDataTuple) -> Optional[SettlementDataTuple]:
-  ker_region = settlement.key[0].lower()
-  ker_municipality = settlement.key[1].lower()
-  ker_settlement = settlement.key[2].lower()
-
-  data_dict = settlement.data
-  result = None
-  result_list = []
-
-  for code, names_list in data_dict.items():
-    for name_data in names_list:
-      region_name = name_data.name[0].lower()
-      municipality_name = name_data.name[1].lower()
-      settlement_name = name_data.name[2].lower()
-
-      straight_case = region_name.find(ker_region) != -1
-
-      # HACK!!! used as a workaround for broken data
-      hack_sf = all([
-        ker_region == 'софийска',
-        region_name.find('софия') != -1
-      ])
-
-      hack_sm = all([
-        ker_region == 'смолян',
-        region_name.find('пловдивска') != -1
-      ])
-
-      hack_pa = all([
-        ker_region == 'пазарджик',
-        region_name.find('пазарджишки') != -1
-      ])
-
-      region_match = any([
-        straight_case,
-        hack_sf,
-        hack_sm,
-        hack_pa
-      ])
-
-      municipality_and_settlement_match = all([
-        municipality_name.find(ker_municipality) != -1,
-        settlement_name.split('.')[-1].strip() == ker_settlement
-      ])
-
-      if all([
-              region_match,
-              municipality_and_settlement_match
-         ]):
-        result_list.append((name_data.end, SettlementDataTuple(settlement.key, code)))
-
-  # if there are multiple matching names take the most recent one
-  sorted(result_list)
-  if len(result_list) > 0:
-    result = result_list[-1][1]
-
-  return result
-
-
-"""## Table Parsing Pipeline"""
-
-
-def fetch_raw_table(data_tuple: DataTuple) -> DataTuple:
-  url = data_tuple.data
-  req = fetch_raw_data(url)
-
-  return DataTuple(req, data_tuple.header_type, data_tuple.table_type)
-
-
-def raw_table_to_lines(data_tuple: DataTuple) -> DataTuple:
-  req = data_tuple.data
-  soup_str: str = str(BeautifulSoup(req.text, 'lxml').prettify(encoding=None))
-  separator = str('\r\n')
-  split = soup_str.split(sep=separator)
-
-  return DataTuple(split, data_tuple.header_type, data_tuple.table_type)
-
-
-def parse_data_line(line: str, table_type: TableTypeEnum) -> Optional[SettlementInfo]:
-  settlement_info_re = ''
-  permanent_population_position = 2
-  current_population_position = -1
-
-  if table_type == TableTypeEnum.Quarterly:
-    settlement_info_re = RegexPatternWrapper().settlement_info_quarterly
-    current_population_position = 3
-  elif table_type == TableTypeEnum.Yearly:
-    settlement_info_re = RegexPatternWrapper().settlement_info_yearly
-    current_population_position = 6
-
-  settlement_info = None
-  if settlement_info_group := regex.search(settlement_info_re, line):
-    name: str = settlement_info_group.group(1)
-    permanent: int = int(settlement_info_group.group(permanent_population_position))
-    current: int = int(settlement_info_group.group(current_population_position))
-
-    name_parts = name.split('.')
-    name = '. '.join([name_parts[0], fix_names(name_parts[1])])
-    settlement_info = SettlementInfo(name.strip(), permanent, current)
-
-  return settlement_info
-
-
-def parse_header_line(
-  line: str,
-  header_type: HeaderEnum,
-  old_header_state: Optional[T]
-) -> Tuple[Optional[MunicipalityIdentifier], Optional[T]]:
-  region_name = None
-
-  if header_type == HeaderEnum.New and (region_gr := regex.search(RegexPatternWrapper().region_name_new, line)):
-    region: str = region_gr.group(1)
-    municipality: str = region_gr.group(2)
-    region_name = MunicipalityIdentifier(region.strip(), municipality.strip())
-
-  elif header_type == HeaderEnum.Old:
-    if old_header_state is None:
-      old_header_state = regex.search(RegexPatternWrapper().old_reg, line)
-      region_name = None
-    elif mun_gr := regex.search(RegexPatternWrapper().old_mun, line):
-      region: str = old_header_state.group(1)
-      municipality: str = mun_gr.group(1)
-      region_name = MunicipalityIdentifier(fix_names(region.strip()), fix_names(municipality.strip()))
-    else:
-      old_header_state = None
-
-  return (region_name, old_header_state)
-
-
-def parse_lines(data_tuple: DataTuple) -> DataTuple:
-  municipality_ids = {}
-  settlements_info = {}
-  old_header_state = None
-
-  for line_num, line in enumerate(data_tuple.data):
-    municipality_id, old_header_state = parse_header_line(line, data_tuple.header_type, old_header_state)
-    if municipality_id:
-      municipality_ids[line_num] = municipality_id
-      continue
-
-    settlement_info = parse_data_line(line, data_tuple.table_type)
-    if settlement_info:
-      settlements_info[line_num] = settlement_info
-
-  return DataTuple(ParsedLines(municipality_ids, settlements_info), data_tuple.header_type, data_tuple.table_type)
-
-
-def parssed_lines_to_full_info_list(data_tuple: DataTuple) -> DataTuple:
-
-  regions = data_tuple.data.municipality_ids
-  settlements_info = data_tuple.data.settlements_info
-
-  reg_keys = list(regions.keys())
-  settlement_keys = list(settlements_info.keys())
-
-  reg_keys_pairs = zip(reg_keys[:-1], reg_keys[1:])
-
-  sk_index = 0
-  full_name_settlement_infos = []
-
-  for current_mun, next_mun in reg_keys_pairs:
-    while current_mun < settlement_keys[sk_index] < next_mun:
-      reg = regions[current_mun]
-      set_info = settlements_info[settlement_keys[sk_index]]
-      fnsi = FullSettlementInfo(fix_names(reg.region),
-                                fix_names(reg.municipality),
-                                fix_names(set_info.name),
-                                set_info.permanent_residents,
-                                set_info.current_residents)
-      full_name_settlement_infos.append(fnsi)
-
-      sk_index += 1
-
-  return DataTuple(full_name_settlement_infos, data_tuple.header_type, data_tuple.table_type)
-
-
-def full_info_list_to_data_frame(data_tuple: DataTuple) -> DataTuple:
-  df = pd.DataFrame(data_tuple.data)
-  df.set_index(['region', 'municipality', 'settlement'], drop=True, inplace=True)
-
-  return DataTuple(df, data_tuple.header_type, data_tuple.table_type)
-
-
-"""## Data Processing Pipeline"""
-
-
-def process_data_tuple(input_data: Tuple[Callable[[DataTuple], DataTuple], DataTuple]) -> DataTuple:
-  parsing_pipeline, data_tuple = input_data
-
-  if data_tuple.table_type == TableTypeEnum.Quarterly:
-    date_group = RegexPatternWrapper().date_group
-  else:
-    date_group = RegexPatternWrapper().year_group
-
-  date_string: str = regex.search(date_group, data_tuple.data).group(1).replace('-', '_')
-  data_frame = parsing_pipeline(data_tuple).data
-  data_frame = data_frame.rename(columns={'permanent_residents': f'permanent_{date_string}',
-                                          'current_residents': f'current_{date_string}'})
-
-  return DataTuple(data_frame, data_tuple.header_type, data_tuple.table_type)
-
-
-def process_data(data_source: List[DataTuple], config: Configuration) -> List[DataTuple]:
-  parsing_pipeline = config['table_parsing']
-  wrapped_data_source = ((parsing_pipeline, dt) for dt in data_source)
-
-  data_frame_list = execute_in_parallel(process_data_tuple, wrapped_data_source)
-
-  if data_frame_list is None:
-    raise Exception('Failed parsing tables!')
-
-  PickleWrapper.pickle_data(data_frame_list, 'data_frames_list')
-
-  return data_frame_list
-
-
-def load_ekatte_dicts() -> Tuple[Dict[Any, Any], Dict[Any, Any]]:
-  processed_sdts = PickleWrapper.load_data('triple_to_ekatte')
-  if (processed_sdts is None) or (not isinstance(processed_sdts, dict)):
-    processed_sdts = {}
-
-  reverse_dict = PickleWrapper.load_data('ekatte_to_triple')
-  if (reverse_dict is None) or (not isinstance(reverse_dict, dict)):
-    reverse_dict = {}
-
-  return processed_sdts, reverse_dict
-
-
-def make_setllements_data_tuple_list(data_frame_list: List[DataTuple]) -> List[Tuple[SettlementDataTuple, str]]:
-  sdt_list = list(map(lambda tup: (SettlementDataTuple(tup[0], tup[0][2]), tup[1]),
-                      map(lambda name: ((fix_names(name[0].strip()),
-                                         fix_names(name[1].strip()),
-                                         fix_names(name[2].split('.')[1].strip())),
-                                        name),
-                          set(chain.from_iterable(
-                              map(lambda dt: dt.data.index.values.tolist(),
-                                  data_frame_list))))))
-  return sdt_list
-
-
-def sleep_time_generator(random_seed: float) -> Generator[float, None, None]:
-  return (st + (st + 1) * random_seed for st in range(round(random_seed), 60, round(5 + 10 * random_seed)))
-
-
-def try_disambiguation(
-  input_data: Tuple[Callable[[SettlementDataTuple], SettlementDataTuple], SettlementDataTuple]
-) -> Tuple[Optional[SettlementDataTuple], SettlementDataTuple]:
-  disambiguation_pipeline, sdt = input_data
-
-  for sleep_time in sleep_time_generator(random.random()):
-    time.sleep(sleep_time)
-    try:
-      result = disambiguation_pipeline(sdt)
-    except ValueError:
-      print(f'Failed disambiguating {sdt} with {sleep_time:.3f}s sleep')
-      continue
-
-    return (result, sdt)
-
-  return (None, sdt)
-
-
-def check_sdt_availability(key: SettlementDataTuple, processed_sdts: Dict[Any, Any], reverse_dict: Dict[Any, Any]) -> bool:
-  return key.key in processed_sdts and processed_sdts[key.key] in reverse_dict
-
-
-def update_data_frame(input_data: Tuple[DataTuple, Dict[Any, Any]]) -> DataTuple:
-  dt, processed_sdts = input_data
-  df = dt.data
-  df.reset_index(inplace=True)
-  df['ekatte'] = df['settlement']
-  cols = df.columns
-
-  def updt(x: Tuple[str, str, str, int, int, str]) -> Tuple[str, str, str, int, int, Optional[str]]:
-    result = (x[0],
-              x[1],
-              x[2],
-              x[3],
-              x[4],
-              processed_sdts.get((fix_names(x[0].strip()),
-                                  fix_names(x[1].strip()),
-                                  fix_names(x[2].split('.')[1].strip())),
-                                 None))
-
-    return result
-
-  df = pd.DataFrame([updt(x) for x in df.to_numpy()])
-
-  df.columns = cols
-  df.dropna(inplace=True)
-  df.set_index(['ekatte'], drop=True, inplace=True)
-
-  df = df.loc[~df.index.duplicated(keep='first')]
-
-  return DataTuple(df, dt.header_type, dt.table_type)
-
-
-def disambiguate_data(data_frame_list: List[DataTuple], config: Configuration) -> List[DataTuple]:
-
-  settlement_disambiguation_pipeline = config['settlement_disambiguation']
-
-  processed_sdts, reverse_dict = load_ekatte_dicts()
-
-  sdt_list = make_setllements_data_tuple_list(data_frame_list)
-  failiures = set()
-
-  wrapped_data_source = ((settlement_disambiguation_pipeline, sdt[0])
-                         for sdt in sdt_list if not check_sdt_availability(sdt[0], processed_sdts, reverse_dict))
-
-  # Higher number of concurrent jobs leads to issues with failing request to NSIs website
-  results = execute_in_parallel(try_disambiguation, wrapped_data_source, 2)
-
-  if results is None:
-    raise Exception('Settlement disambiguation failed!')
-
-  for value, sdt in results:
-    if value is None:
-      failiures.add(sdt[0])
-      PickleWrapper.pickle_data(failiures, 'failiures')
-    else:
-      processed_sdts[value.key] = value.data
-      PickleWrapper.pickle_data(processed_sdts, 'triple_to_ekatte')
-
-      reverse_dict[value.data] = sdt[1]
-      PickleWrapper.pickle_data(reverse_dict, 'ekatte_to_triple')
-
-  wrapped_data_tuple_source = ((dt, processed_sdts) for dt in data_frame_list)
-  disambiguated_data = execute_in_parallel(update_data_frame, wrapped_data_tuple_source)
-
-  if disambiguated_data is None:
-    raise Exception('Updating DataFrames failed!')
-
-  PickleWrapper.pickle_data(disambiguated_data, 'data_frames_list_disambiguated')
-
-  return disambiguated_data
-
-
-def combine_data(processed_data: List[DataTuple], config: Configuration) -> List[DataTuple]:
-  combined: Optional[pd.DataFrame] = None
-  names = ['region', 'municipality', 'settlement']
-
-  for dt in processed_data:
-    if combined is None:
-      combined = dt.data.drop(labels=names, axis=1)
-    else:
-      df: pd.DataFrame = dt.data.drop(labels=names, axis=1)
-      combined = combined.merge(df,
-                                how='outer',
-                                copy=False,
-                                left_index=True,
-                                right_index=True)
-
-  if combined is None:
-    raise Exception('Failed to combine DataFarmes')
-
-  combined.fillna(value=0, inplace=True)
-  for column in combined.columns.to_list():
-    combined[column] = combined[column].astype(int)
-
-  PickleWrapper.pickle_data(combined, 'combined_tables')
-
-  return [DataTuple(combined, HeaderEnum(0), TableTypeEnum(0))]
-
-
-def store_data_list(processed_data: List[DataTuple], config: Configuration) -> List[DataTuple]:
-  for dt in processed_data:
-    df: pd.DataFrame = dt.data
-
-    name = f'grao_data_{"_".join(df.columns[-1].split("_")[1:])}'
-    df.to_csv(f'{config.processed_tables_path}/{name}.csv')
-
-  return processed_data
-
-
-def store_combined_data(processed_data: List[DataTuple], config: Configuration) -> List[DataTuple]:
-  combined_data: pd.DataFrame = processed_data[0].data
-
-  combined_data.to_csv(f'{config.combined_tables_path}/grao_data_combined.csv')
-
-  return processed_data
-
-
-"""## Visualizations"""
-
-
-def autolabel(ax, rects: Iterable[Any]):
-    """Attach a text label above each bar in *rects*, displaying its height."""
-    for rect in rects:
-        height = rect.get_height()
-        ax.annotate('{}'.format(height),
-                    xy=(rect.get_x() + rect.get_width() / 2, height),
-                    xytext=(0, 3),  # 3 points vertical offset
-                    textcoords="offset points",
-                    ha='center', va='bottom')
-
-
-def load_processed_data() -> Tuple[Dict[Any, Any], Dict[Any, Any]]:
-  ekatte_to_triple = PickleWrapper.load_data('ekatte_to_triple')
-  combined = PickleWrapper.load_data('combined_tables')
-
-  if ekatte_to_triple is None or combined is None:
-    raise Exception('There was an issue loading the data!')
-
-  combined_dict = combined.to_dict(orient='index')
-
-  return ekatte_to_triple, combined_dict
-
-
-def path_for_settlement_graphic(directory: str, name: str, sufix: str = '') -> str:
-  modified_name = name.replace('.', '').replace(' ', '_')
-
-  if sufix == '_':
-    sufix = ''
-
-  return f'{directory}/{modified_name}{sufix}'
-
-
-def prepare_directory(triple: Tuple[str, str, str], base: str) -> str:
-  sub_path = f'{triple[0]}/{triple[1]}'.replace(' ', '_')
-  full_path = f'{base}/{sub_path}'
-
-  if not os.path.exists(full_path):
-      os.makedirs(full_path)
-
-  return full_path
-
-
-def clear_figure():
-  # Clear the current axes.
-  plt.cla()
-
-  # Clear the current figure.
-  plt.clf()
-  plt.ioff()
-
-
-def draw_plot(directory: str, plot_name: str, type_name: str):
-  plt_path = path_for_settlement_graphic(directory, plot_name.split(',')[-1].strip(), f'_{type_name.lower()}')
-  plt.savefig(plt_path)
-  print(plt_path)
-  clear_figure()
-
-
-def plot_single_value(directory: str, plot_name: str, values: List[int], labels: List[str], type_name: str = ''):
-  _, ax = plt.subplots()
-
-  xticks = np.arange(len(labels))
-  width = 0.4
-
-  # Add some text for labels, title and custom x-axis tick labels, etc.
-  ax.set_ylabel('Number of residents')
-  ax.set_xticks(xticks)
-  ax.set_xticklabels(labels)
-  ax.set_title(plot_name)
-
-  rects = ax.bar(xticks - width / 20, values, width, label=type_name.capitalize(), align='center')
-  autolabel(ax, rects)
-
-  ax.legend()
-
-  draw_plot(directory, plot_name, type_name)
-
-
-def plot_comparison(
-  directory: str,
-  settlement_name: str,
-  values_list: List[List[int]],
-  labels: List[str],
-  type_name: str = ''
-):
-  _, ax = plt.subplots()
-  xticks = np.arange(len(labels))
-
-  ax.set_xticks(xticks)
-  ax.set_xticklabels(labels)
-
-  for values in values_list:
-    plt.plot(values)
-
-  plt.title('Comparison between permanent and current')
-  plt.xlabel('Year')
-  plt.ylabel('Number of residents')
-  plt.legend(['Permanent', 'Current'], loc='upper right')
-
-  draw_plot(directory, settlement_name, type_name)
-
-
-def create_visualizations(config: Configuration):
-  ekatte_to_triple, combined_dict = load_processed_data()
-
-  plt.rcParams['figure.figsize'] = [45, 15]
-
-  labels = list(combined_dict[list(combined_dict.keys())[0]].keys())
-  date_labels = list(map(lambda l: ' '.join(l.split('_')[1:]), labels[0::2]))
-  date_labels.reverse()
-
-  for item in combined_dict:
-    triple = ekatte_to_triple[item]
-    name = f'обл. {triple[0]}, общ. {triple[1]}, {triple[2]}'
-    full_path = prepare_directory(triple, config.visualizations_path)
-
-    values = list(combined_dict[item].values())
-
-    permanent_values = values[0::2]
-    permanent_values.reverse()
-    plot_single_value(full_path, name, permanent_values, date_labels, 'permanent')
-
-    current_values = values[1::2]
-    current_values.reverse()
-    plot_single_value(full_path, name, current_values, date_labels, 'current')
-
-    plot_comparison(full_path, name, [permanent_values, current_values], date_labels, 'compare')
-
-    # Closes all the figure windows.
-    plt.close('all')
-
-
-"""## Matching data with QID"""
-
-
-def find_ref_url(path_to_file: str, file_prefix: str, url_list: List[str]) -> str:
-  processing_pipline = Pipeline((
-    (lambda path: os.path.basename(path)),
-    (lambda name: name.split('.')[0]),
-    (lambda name: name.replace(file_prefix, '')),
-    (lambda name: name.replace('_', '-')),
-    (lambda date_str: next(filter((lambda url: url.find(date_str) > -1), url_list))),
-  ))
-
-  result = processing_pipline(path_to_file)
-  return result
-
-
-def date_from_url(url: str) -> datetime.datetime:
-  date_str: str = ''
-
-  if date_group := regex.search(RegexPatternWrapper().full_date_group, url):
-    date_str = date_group.group(1)
-  elif date_group := regex.search(RegexPatternWrapper().year_group, url):
-    date_str = date_group.group(1)
-    date_str = f'31-12-{date_str}'
-
-  date = datetime.datetime.strptime(date_str, '%d-%m-%Y')
-
-  return date
-
-
-def file_prefix_for_directory(directory: str):
-  return f'{os.path.basename(directory)}_'
-
-
-def find_date_sufix(url: str) -> str:
-    date = date_from_url(url)
-    date_sufix = f'{date.year}'
-
-    if date.day != 31 and date.month != 12:
-      date_sufix = f'{date.month:02}_{date_sufix}'
-
-    return date_sufix
-
-
-def single_processed_file_info(input_data: Tuple[str, str, List[str]]) -> Tuple[datetime.datetime, str, str]:
-  file, storage_directory, url_list = input_data
-  file_prefix = file_prefix_for_directory(storage_directory)
-
-  url = find_ref_url(file, file_prefix, url_list)
-  date = date_from_url(url)
-
-  return (date, url, os.path.join(storage_directory, file))
-
-
-def find_latest_processed_file_info(storage_directory: str, url_list: List[str]) -> Tuple[datetime.datetime, str, str]:
-  wrapped_data_generator = ((file, storage_directory, url_list) for file in os.listdir(storage_directory))
-
-  processed_files = execute_in_parallel(single_processed_file_info, wrapped_data_generator)
-
-  if processed_files is None:
-    raise Exception(f'Couldn\'t processed files in {storage_directory}')
-
-  processed_files = sorted(processed_files)
-  return processed_files[-1]
-
-
-def dict_from_csv(csv_path: str, index_name: str) -> Dict[Any, Any]:
-  loaded_dict = pd.DataFrame(pd.read_csv(csv_path, dtype=np.str)).set_index(index_name).to_dict(orient='index', into=dict)
-
-  if not isinstance(loaded_dict, dict):
-    result: Dict[Any, Any] = {}
-  else:
-    result = loaded_dict
-
-  return result
-
-
-def update_matched_data(config: Configuration):
-  matched_tables_path = config.matched_tables_path
-  matched_data_time, _, matched_data_path = find_latest_processed_file_info(
-    matched_tables_path,
-    config.data
-  )
-  grao_data_time, grao_data_url, grao_data_path = find_latest_processed_file_info(
-    config.processed_tables_path,
-    config.data
-  )
-
-  if grao_data_time <= matched_data_time:
-    return
-
-  date_sufix = find_date_sufix(grao_data_url)
-
-  matched_data_dict = dict_from_csv(matched_data_path, index_name='ekatte')
-  grao_data_dict = dict_from_csv(grao_data_path, index_name='ekatte')
-
-  new_matched_data = {}
-  for key, value in matched_data_dict.items():
-    new_matched_data[key] = value
-    new_matched_data[key]['permanent_population'] = grao_data_dict[key][f'permanent_{date_sufix}']
-    new_matched_data[key]['current_population'] = grao_data_dict[key][f'current_{date_sufix}']
-
-  new_matched_df = pd.DataFrame.from_dict(new_matched_data, orient='index', dtype=np.str).reset_index()
-
-  if new_matched_df is None:
-    raise Exception('Failed to cerate updated DataFrame')
-
-  new_matched_df.rename(columns={'index': 'ekatte'}, inplace=True)
-  new_matched_df.to_csv(f'{matched_tables_path}/{file_prefix_for_directory(matched_tables_path)}{date_sufix}.csv', index=False)
-
-
-"""## Update bot"""
-
-
-def create_qualifiers(date: datetime.datetime) -> List[Any]:
-  ref_time_str = f'+{date.isoformat()}Z'
-  point_in_time = wdi_core.WDTime(time=ref_time_str, prop_nr='P585', is_qualifier=True)
-  determination_method = wdi_core.WDItemID(value='Q90878157', prop_nr="P459", is_qualifier=True)
-
-  return [point_in_time, determination_method]
-
-
-def login_with_credentials(credentials_path: str) -> wdi_login.WDLogin:
-  credentials = pd.DataFrame(pd.read_csv(credentials_path))
-  username, password = tuple(credentials)
-
-  return wdi_login.WDLogin(username, password)
-
-
-def update_item(login: wdi_login.WDLogin, settlement_qid: str, data: List[wdi_core.WDQuantity]):
-  item = wdi_core.WDItemEngine(wd_item_id=settlement_qid, data=data)
-  item.write(login, False)
-  time.sleep(15)
-
-
-def update_all_settlements(config: Configuration):
-  login = login_with_credentials(config.credentials_path)
-
-  ref_time, ref_url, path = find_latest_processed_file_info(config.matched_tables_path, config.data)
-
-  ref = wdi_core.WDUrl(prop_nr="P854", value=ref_url, is_reference=True)
-  # publisher = wdi_core.WDItemID(value=login.consumer_key, prop_nr="P123", is_reference=True)
-
-  qualifiers = create_qualifiers(ref_time)
-
-  error_logs = []
-
-  data = pd.DataFrame(pd.read_csv(path))
-  for _, row in data.iterrows():
-    settlement_qid: str = row['settlement']
-    population: str = row['permanent_population']
-    prop = wdi_core.WDQuantity(
-      prop_nr='P1082',
-      value=population,
-      qualifiers=qualifiers,
-      references=[[ref]]
-    )
-
-    try:
-      update_item(login, settlement_qid, [prop])
-    except BaseException:
-      error_logs.append(settlement_qid)
-      print("An error occured for item : " + settlement_qid)
-
-  if len(error_logs) > 0:
-    print("Summarizing failures for specific IDs")
-    for error in error_logs:
-      print("Error for : " + error)
+from typing import Callable, List, Optional
+from dataclasses import dataclass
+
+from grao_tables_processing.common.custom_types import T, U
+from grao_tables_processing import Configuration
+from grao_tables_processing import PickleWrapper
+
+from grao_tables_processing import settlement_disambiguation
+from grao_tables_processing import table_parser
+from grao_tables_processing import create_table_processor
+from grao_tables_processing import create_visualizations
+from grao_tables_processing import update_matched_data, update_all_settlements
 
 
 """## Input validation """
@@ -1121,7 +92,7 @@ def main():
   """
 
   parser = argparse.ArgumentParser(description="Processes the tables provided by GRAO and"
-                                               "extracts th einformation from them to csv files",
+                                               "extracts the information from them to csv files",
                                    epilog=example_text,
                                    formatter_class=argparse.RawDescriptionHelpFormatter)
   parser.add_argument("--data_configuration_path",
@@ -1147,7 +118,7 @@ def main():
                       help="Path to the file containing credentials.")
   parser.add_argument("--produce_graphics",
                       default=False, action="store_true",
-                      help="If set the script will produce grafics from the processed tables.")
+                      help="If set the script will produce graphics from the processed tables.")
   parser.add_argument("--update_wiki_data",
                       default=False, action="store_true",
                       help="If set the script will update WikiData with the processed tables.")
@@ -1193,30 +164,10 @@ def main():
     args.credentials_path
   )
 
-  settlement_disambiguation = Pipeline(functions=(
-    fetch_raw_settlement_data,
-    parse_raw_settlement_data,
-    mach_key_with_code
-  ))
   configuration['settlement_disambiguation'] = settlement_disambiguation
+  configuration['table_parser'] = table_parser
 
-  table_parsing = Pipeline(functions=(
-    fetch_raw_table,
-    raw_table_to_lines,
-    parse_lines,
-    parssed_lines_to_full_info_list,
-    full_info_list_to_data_frame,
-  ))
-  configuration['table_parsing'] = table_parsing
-
-  processing_pipeline = Pipeline(functions=(
-    (lambda data: process_data(data, configuration)),
-    (lambda data: disambiguate_data(data, configuration)),
-    (lambda data: store_data_list(data, configuration)),
-    (lambda data: combine_data(data, configuration)),
-    (lambda data: store_combined_data(data, configuration)),
-  ))
-
+  processing_pipeline = create_table_processor(configuration)
   data_source = configuration.process_data_configuration()
 
   processing_pipeline(data_source)
