+++ grao_tables_processing/common/singleton.py
@@ -1,0 +1,10 @@
+from typing import Any, List, Dict
+
+
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args: List[Any], **kwargs: Dict[str, Any]):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
