+++ grao_tables_processing/visualization/__init__.py
@@ -1,0 +1,3 @@
+import grao_tables_processing.visualization.visualization as v
+
+create_visualizations = v.create_visualizations
